Add support for repeatable, numbered board layouts
This feature allows the bot developer to note and return to a given
board layout based on a board number. The default behavior is identical
to the production package except for the presence of a Board Number
field below the board. If a given board is of interest for later study,
the number can be noted and reentered.

- Incorporated David Bau's seedrandom.js since Javascript's Math.random()
  doesn't provide a means for seeding. Take note of the comments in
  board.js about why multiple random number generators are used.

- Allows the developer to provide a default_board_number() function in
  their bot script to control what board will be displayed.